fix add auth to pdfs route
